make all inc* arguments as type int

+++ src/gpuarray_blas_opencl_clblas.c
@@ -146,8 +146,8 @@ static int dgemmBatch(cb_order order, cb_transpose transA, cb_transpose transB,
 static int hgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, float alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, size_t incX,
-                      float beta, gpudata **y, size_t *offY, size_t incY,
+                      gpudata **x, size_t *offX, int incX,
+                      float beta, gpudata **y, size_t *offY, int incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
@@ -155,8 +155,8 @@ static int hgemvBatch(cb_order order, cb_transpose transA,
 static int sgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, float alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, size_t incX,
-                      float beta, gpudata **y, size_t *offY, size_t incY,
+                      gpudata **x, size_t *offX, int incX,
+                      float beta, gpudata **y, size_t *offY, int incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
@@ -164,31 +164,31 @@ static int sgemvBatch(cb_order order, cb_transpose transA,
 static int dgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, double alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, size_t incX,
-                      double beta, gpudata **y, size_t *offY, size_t incY,
+                      gpudata **x, size_t *offX, int incX,
+                      double beta, gpudata **y, size_t *offY, int incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int hgerBatch(cb_order order, size_t M, size_t N, float alpha,
-                     gpudata **x, size_t *offX, size_t incX,
-                     gpudata **y, size_t *offY, size_t incY,
+                     gpudata **x, size_t *offX, int incX,
+                     gpudata **y, size_t *offY, int incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sgerBatch(cb_order order, size_t M, size_t N, float alpha,
-                     gpudata **x, size_t *offX, size_t incX,
-                     gpudata **y, size_t *offY, size_t incY,
+                     gpudata **x, size_t *offX, int incX,
+                     gpudata **y, size_t *offY, int incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int dgerBatch(cb_order order, size_t M, size_t N, double alpha,
-                     gpudata **x, size_t *offX, size_t incX,
-                     gpudata **y, size_t *offY, size_t incY,
+                     gpudata **x, size_t *offX, int incX,
+                     gpudata **y, size_t *offY, int incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
@@ -196,24 +196,24 @@ static int dgerBatch(cb_order order, size_t M, size_t N, double alpha,
 
 static int hdot(
         size_t N,
-        gpudata *X, size_t offX, size_t incX,
-        gpudata *Y, size_t offY, size_t incY,
+        gpudata *X, size_t offX, int incX,
+        gpudata *Y, size_t offY, int incY,
         gpudata *Z, size_t offZ) {
     return GA_DEVSUP_ERROR;
 }
 
 static int sdot(
         size_t N,
-        gpudata *X, size_t offX, size_t incX,
-        gpudata *Y, size_t offY, size_t incY,
+        gpudata *X, size_t offX, int incX,
+        gpudata *Y, size_t offY, int incY,
         gpudata *Z, size_t offZ) {
     return GA_DEVSUP_ERROR;
 }
 
 static int ddot(
         size_t N,
-        gpudata *X, size_t offX, size_t incX,
-        gpudata *Y, size_t offY, size_t incY,
+        gpudata *X, size_t offX, int incX,
+        gpudata *Y, size_t offY, int incY,
         gpudata *Z, size_t offZ) {
     return GA_DEVSUP_ERROR;
 }
